Renamed and relocated record for DCs damage distribution
